Fix warning

+++ Direct3DBaseProject/ButtonA.cpp
@@ -34,7 +34,7 @@ ButtonA::ButtonA():
 	m_origin.x = float(texDesc.Width / 2);
 	m_origin.y = float(texDesc.Height / 2);
 
-	m_pos.x = deviceAccessor->GetScreenSize().right / 2;
+	m_pos.x = (float)deviceAccessor->GetScreenSize().right / 2;
 	m_pos.y = float(json->GetData()["ButtonAPosition"].at(1));
 }
