+++ Direct3DBaseProject/GameOverUI.cpp
@@ -34,7 +34,7 @@ GameOverUI::GameOverUI():
 	m_origin.x = float(texDesc.Width / 2);
 	m_origin.y = float(texDesc.Height / 2);
 
-	m_pos.x = deviceAccessor->GetScreenSize().right / 2;
+	m_pos.x = (float)deviceAccessor->GetScreenSize().right / 2;
 	m_pos.y = float(json->GetData()["GameOverUIPosY"]);
 }
